Add zero initializers to BrigDirectiveVariableInit when necessary
The HSA PRM 19.6.19 says,

If the symbol is an array with a size, then dim must be the number of elements
in the array. If the symbol is not an array or is a flexible array, then dim
must be 0. An array declared in the textual form without a size, but with an
initializer, is not considered a flexible array. In this case, the value of dim
is related to the directive referenced by the init field: if
BrigDirectiveVariableInit, then dim must be the same value as elementCount field
of the BrigDirectiveVariableInit; if BrigDirectiveImageInit or
BrigDirectiveSamplerInit, then dim must be 1 because HSAIL allows only a single
initializer for image and sampler objects.

But for the example:
global_u32 &array[10] = {1,2,3,4,5,6,7,8};

hsailasm sets dim to 10 and elementCount to 8.

Regarding arrays, HSA PRM 4.19.1 says,

If there is no initializer, the value of the object is undefined. However, if
some of the elements are initialized, then the rest are initialized to zero.