Prototype of test framework
In order to ensure that libHSAIL is working properly, especially
on the port to Linux, there is a need of a test framework that
is capable on checking that the generated assembly/disassembly
is correct.

This commit adds an implementation of such framework. It still
requires some code cleanup before considering merging with master.