Align the debug sections to 4-bytes.
The Elf spec requires that relocation sections be aligned to sizeof(Elf32_Sym)
for 32-bit Elf files. Not aligning the sections caused problems with LLVM code
that asserted that the sections were properly aligned. In particular, it
caused almost all test cases in brig_reader_test to fail when compiled in
debug mode on AMD64 with recent releases of clang.