Many cmake fixes
The primary purpose is to stop hardcoding paths
in OS dependent shell scripts which is the point of
using cmake. Find the correct dependencies using the
search paths which is what should be specified.

Also fixes other issues, such as redundant lib prefixes
on the final libraries, so liblibhsail is no longer built.
Use the current directory relative paths instead of the
cmake root path so that embedding the build in other build
systems hopefully works now. Emit generated files to the
build directory instead of the source.