Fix for issue #52
Multiplication has higher precedence than ternary operator. So `a[0] * numElems` is computed first and used as the 1st operand of `?:`. This seems semantically wrong and also causes the warning (because the 1st operand of `?:` has boolean context). Parntheses around "a[0]" are not actually required, but with these the code looks clearer to me.